requirements.txt
joblib==1.1.0
nltk==3.6.5
numpy==1.21.4
opencv_python_headless==4.5.5.64
pandas
Pillow==9.1.1
streamlit==1.9.0
tensorflow==2.9.0
tf_nightly_gpu
